+++ src/dependency_injector/errors.py
@@ -6,3 +6,7 @@ class Error(Exception):
 
     All dependency injector errors extend this error class.
     """
+
+
+class NoSuchProviderError(Error, AttributeError):
+    """Error that is raised when provider lookup is failed."""