+++ src/dependency_injector/providers.pyx
@@ -9,7 +9,10 @@ import sys
 import types
 import threading
 
-from .errors import Error
+from .errors import (
+    Error,
+    NoSuchProviderError,
+)
 
 cimport cython
 
@@ -1109,6 +1112,81 @@ cdef class FactoryDelegate(Delegate):
             raise Error('{0} can wrap only {1} providers'.format(
                 self.__class__, Factory))
         super(Delegate, self).__init__(factory)
+
+
+cdef class FactoryAggregate(Provider):
+    """Factory providers aggregate.
+
+    :py:class:`FactoryAggregate` is an aggregate of :py:class:`Factory`
+    providers.
+
+    :py:class:`FactoryAggregate` is a delegated provider, meaning that it is
+    injected "as is".
+
+    All aggregated factories could be retrieved as a read-only
+    dictionary :py:attr:`FactoryAggregate.factories` or just as an attribute of
+    :py:class:`FactoryAggregate`.
+    """
+
+    __IS_DELEGATED__ = True
+
+    def __init__(self, **factories):
+        """Initializer.
+
+        :param factories: Dictionary of aggregate factories.
+        :type factories: dict[str, :py:class:`Factory`]
+        """
+        for factory in factories.values():
+            if isinstance(factory, Factory) is False:
+                raise Error(
+                    '{0} can aggregate only instances of {1}, given - {2}'
+                    .format(self.__class__, Factory, factory))
+        self.__factories = factories
+        super(FactoryAggregate, self).__init__()
+
+    def __call__(self, factory_name, *args, **kwargs):
+        """Create new object using factory with provided name.
+
+        Callable interface implementation.
+        """
+        return self.__get_factory(factory_name)(*args, **kwargs)
+
+    def __getattr__(self, factory_name):
+        """Return aggregated factory."""
+        return self.__get_factory(factory_name)
+
+    def __str__(self):
+        """Return string representation of provider.
+
+        :rtype: str
+        """
+        return represent_provider(provider=self, provides=self.factories)
+
+    @property
+    def factories(self):
+        """Return dictionary of factories, read-only."""
+        return self.__factories
+
+    def override(self, _):
+        """Override provider with another provider.
+
+        :param provider: Overriding provider.
+        :type provider: :py:class:`Provider`
+
+        :raise: :py:exc:`dependency_injector.errors.Error`
+
+        :return: Overriding context.
+        :rtype: :py:class:`OverridingContext`
+        """
+        raise Error(
+            '{0} providers could not be overridden'.format(self.__class__))
+
+    cdef Factory __get_factory(self, str factory_name):
+        if factory_name not in self.__factories:
+            raise NoSuchProviderError(
+                '{0} does not contain factory with name {1}'.format(
+                    self, factory_name))
+        return <Factory> self.__factories[factory_name]
 
 
 cdef class BaseSingleton(Provider):