+++ docs/main/changelog.rst
@@ -9,8 +9,9 @@ follows `Semantic versioning`_
 
 Development version
 -------------------
+- Add ``FactoryAggregate`` provider.
 - Add support of six 1.11.0.
-- Regenerate C sources using Cython 0.27.
+- Regenerate C sources using Cython 0.27.1.
 
 3.6.1
 -----